+++ src/Components/Signup/licenses.js
@@ -1,0 +1,43 @@
+export const licenses = [
+  "None Yet, I'm New!",
+  "USCG Master 100gt",
+  "USCG Master 200gt",
+  "USCG Master 500gt",
+  "USCG Master 1600gt",
+  "USCG Master 1600gt",
+  "USCG Master Unlimited",
+  "STCW Basic Safety Training",
+  "RYA Yachtmaster",
+  "RYA Day Skipper",
+  "RYA Coastal Skipper",
+  "ICC (International Certificate of Competence)",
+  "MCA Master 200gt",
+  "MCA Master 500gt",
+  "MCA Master 3000gt",
+  "MCA Master Unlimited",
+  "MCA OOW (Officer of the Watch)",
+  "MCA Chief Mate 3000gt",
+  "MCA Chief Mate Unlimited",
+  "RYA Powerboat Level 2",
+  "RYA Personal Watercraft (PWC) Instructor",
+  "STCW Basic Training",
+  "STCW Advanced Firefighting",
+  "STCW Proficiency in Survival Craft and Rescue Boats",
+  "STCW Engine Room Rating",
+  "STCW Able Seafarer Engine",
+  "STCW Officer of the Watch (Engine)",
+  "MCA Marine Engine Operator License (MEOL)",
+  "MCA AEC 1",
+  "MCA AEC 2",
+  "MCA Chief Engineer (Y4)",
+  "MCA Chief Engineer (Y3)",
+  "MCA Chief Engineer (Y2)",
+  "MCA Chief Engineer (Y1)",
+  "MCA Second Engineer (500 GT)",
+  "MCA Second Engineer (3000 GT)",
+  "MCA Second Engineer (Unlimited)",
+  "MCA Electro-Technical Officer (ETO)",
+  "USCG Designated Duty Engineer (DDE)",
+  "USCG Assistant Engineer",
+  "USCG Chief Engineer",
+];
