+++ src/pages/index.js
@@ -4,27 +4,27 @@ import Header from "@/Components/Header";
 import Hero from "@/Components/Hero";
 import About from "@/Components/About";
 import Background from "@/Components/Background";
-import Signup from "@/Components/Signup";
+import Signup from "@/Components/Signup/Signup";
 import ContactForm from "@/Components/ContactForm";
 
 const Home = () => {
   return (
     <div className="bg-[#363636] text-white h-screen overflow-scroll">
-      <Background/>
+      <Background />
       <Head>
         <title>CrewBlast Yachting</title>
       </Head>
       <Header />
-      <section id='home'>
-      <Hero />
+      <section id="home">
+        <Hero />
       </section>
       <section id="about">
         <About />
       </section>
-      <section id='signup'>
-        <Signup/>
+      <section id="signup">
+        <Signup />
       </section>
-      <section id='contact'>
+      <section id="contact">
         <ContactForm />
       </section>
     </div>
